+++ mkm/src/main/kotlin/mkm/repos/CardRepository.kt
@@ -7,7 +7,7 @@ import org.springframework.data.repository.query.Param
 import org.springframework.stereotype.Repository
 
 @Repository
-interface CardRepository: CrudRepository<Card, Int> {
+interface CardRepository: CrudRepository<Card, Long> {
     @Query("SELECT c FROM Card c WHERE UPPER(c.englishName) LIKE '%' || UPPER(:englishName) || '%'")
     fun findByEnglishName(@Param("englishName") englishName: String): List<Card>
 }
